Coded many different sections in order to get the drawing to be able to work.
In the shape canvas, it has been made so that you are able to freely draw on the canvas and change the width of the line.
In the DrawingPanel, different listeners and imports have been added so that it can detect mouse movements and such for the program to do things with.